requirements.txt
anthropic
selenium
beautifulsoup4
webdriver-manager
python-dotenv
scikit-learn
sentence-transformers
